DOC-206 rename Versioning Test to check that Ceph-tests runs in travis
