+++ datasafe-examples/datasafe-examples-versioned-s3/src/test/resources/simplelogger.properties
@@ -1,0 +1,9 @@
+org.slf4j.simpleLogger.defaultLogLevel: info
+
+org.slf4j.simpleLogger.log.de.adorsys.datasafe.examples.business.s3.BaseUserOperationsWithDefaultDatasafeOnVersionedStorageTest=debug
+
+org.slf4j.simpleLogger.dateTimeFormat=HH:mm:ss.SSS
+org.slf4j.simpleLogger.showDateTime=true
+org.slf4j.simpleLogger.showThreadName=true
+
+org.slf4j.simpleLogger.logFile=System.out
