+++ datasafe-simple-adapter/datasafe-simple-adapter-api/src/main/java/de/adorsys/datasafe/simple/adapter/api/SimpleDatasafeService.java
@@ -15,18 +15,12 @@ public interface SimpleDatasafeService {
     void destroyUser(UserIDAuth userIDAuth);
 
     boolean userExists(UserID userID);
-
-    void registerDFSCredentials (UserIDAuth userIDAuth, DFSCredentials dfsCredentials);
     /**
      * Document
      */
     void storeDocument(UserIDAuth userIDAuth, DSDocument dsDocument);
 
     DSDocument readDocument(UserIDAuth userIDAuth, DocumentFQN documentFQN);
-
-    void storeDocumentStream(UserIDAuth userIDAuth, DSDocumentStream dsDocumentStream);
-
-    DSDocumentStream readDocumentStream(UserIDAuth userIDAuth, DocumentFQN documentFQN);
 
     void deleteDocument(UserIDAuth userIDAuth, DocumentFQN documentFQN);
 
@@ -35,23 +29,5 @@ public interface SimpleDatasafeService {
     void deleteFolder(UserIDAuth userIDAuth, DocumentDirectoryFQN documentDirectoryFQN);
 
     List<DocumentFQN> list(UserIDAuth userIDAuth, DocumentDirectoryFQN documentDirectoryFQN, ListRecursiveFlag recursiveFlag);
-
-    /**
-     * InboxStuff
-     */
-    List<DocumentFQN> listInbox(UserIDAuth userIDAuth);
-
-    void writeDocumentToInboxOfUser(UserID receiverUserID, DSDocument document, DocumentFQN destDocumentFQN);
-
-    DSDocument readDocumentFromInbox(UserIDAuth userIDAuth, DocumentFQN source);
-
-    void deleteDocumentFromInbox(UserIDAuth userIDAuth, DocumentFQN documentFQN);
-
-    /**
-     * conveniance methods
-     */
-    void moveDocumnetToInboxOfUser(UserIDAuth userIDAuth, UserID receiverUserID, DocumentFQN sourceDocumentFQN, DocumentFQN destDocumentFQN, MoveType moveType);
-
-    DSDocument moveDocumentFromInbox(UserIDAuth userIDAuth, DocumentFQN source, DocumentFQN destination);
 }
