+++ datasafe-simple-adapter/datasafe-simple-adapter-impl/src/test/java/de/adorsys/datasafe/simple/adapter/impl/SimpleDatasafeAdapterTest.java
@@ -82,7 +82,7 @@ public class SimpleDatasafeAdapterTest extends WithStorageProvider {
 
     private static Stream<Arguments> parameterCombination() {
             return Sets.cartesianProduct(
-                    allDefaultStorages().collect(Collectors.toSet()),
+                    allStorages().collect(Collectors.toSet()),
                     withOrWithoutEncryption().collect(Collectors.toSet())
             ).stream().map(it -> Arguments.of(it.get(0), it.get(1)));
     }
